+++ src/code_generator.h
@@ -213,7 +213,7 @@ bool add_null();
  * @param val The float value to be added.
  * @return true if the operation was successful, false otherwise.
  */
-bool add_float(float val);
+bool add_float(double val);
 
 /**
  * @brief Add formatted string value to the code buffer.
